factor config parser into a directory
